pyproject.toml
[project]
name = "ai-resume-analyzer"
version = "0.1.0"
description = "A simple AI Resume Analyzer"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.9.0.post1",
    "langchain-community>=0.3.14",
    "langchain-groq>=0.2.3",
    "langchain-huggingface>=0.1.2",
    "langchain-ollama>=0.2.2",
    "mkdocs-material>=9.5.50",
    "pypdf>=5.1.0",
    "python-dotenv>=1.0.1",
    "streamlit>=1.41.1",
]

[dependency-groups]
dev = [
    "black>=24.10.0",
    "invoke>=2.2.0",
    "ruff>=0.9.2",
]

[project.scripts]
streamlit = "streamlit.cli:main"
